+++ src/domains/MyProfile/Page/index.tsx
@@ -1,20 +1,22 @@
+import { Link } from 'react-router-dom';
+import { useGetProfileQuery } from '@/api/userApi.ts';
+
 import Header from '@/components/common/Layout/Header';
 import Footer from '@/components/common/Layout/Footer';
 import UserContainer from '@/components/common/UserContainer';
-import { Container, Menu } from '@/components/MyProfile/myProfile.style.ts';
 import { HeaderItem } from '@/components/common/Layout/Header/Header.style.ts';
+import LoadingIcon from '@/components/common/LoadingIcon';
+import NoData from '@/components/common/NoData.tsx';
 
 import MyProfileIcon from '@/assets/icons/header/my-porfile-icon.svg?react';
-import { useGetProfileQuery } from '@/api/userApi.ts';
-import MenuItem from '@/components/MyProfile/MenuItem.tsx';
+import MenuItem from '@/domains/MyProfile/components/MenuItem.tsx';
 import FileIcon from '@/assets/icons/myprofile/file-icon.svg?react';
 // import NoticeIcon from '@/assets/icons/notice-icon.svg?react';
 // import Toggle from '@/components/common/Toggle.tsx';
-import LoadingIcon from '@/components/common/LoadingIcon';
-import NoData from '@/components/common/NoData.tsx';
-import { Link } from 'react-router-dom';
 
-const MyProfilePage = () => {
+import { Container, Menu } from './page.style.ts';
+
+const Index = () => {
   const { data, isLoading } = useGetProfileQuery(null);
   if (isLoading) return <LoadingIcon />;
   if (!data) return <NoData>데이터를 찾을 수 없습니다</NoData>;
@@ -47,4 +49,4 @@ const MyProfilePage = () => {
   );
 };
 
-export default MyProfilePage;
+export default Index;
