+++ src/domains/MyProfile/Page/page.style.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  height: 100%;
+`;
+
+export const Menu = styled.ul`
+  display: flex;
+  flex-direction: column;
+
+  padding: 25px 20px;
+`;
