+++ src/domains/MyProfile/components/MenuItem.style.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const Container = styled.li`
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 16px;
+
+  & > div {
+    display: flex;
+    & > svg {
+      margin-right: 8px;
+    }
+  }
+`;
